+++ tests/en_test.go
@@ -11,8 +11,8 @@ const langEn = "en"
 
 func TestParseEn(t *testing.T) {
 	cases := []struct {
-		date   time.Time
-		result string
+		date time.Time
+		res  string
 	}{
 		{subMinutes(1), "1 minute ago"},
 		{subMinutes(2), "2 minutes ago"},
@@ -57,8 +57,14 @@ func TestParseEn(t *testing.T) {
 		t.Run("result for "+tc.date.String(), func(test *testing.T) {
 			timeago.Configure(&timeago.Config{Language: langEn})
 
-			if res := timeago.Parse(tc.date); res != tc.result {
-				test.Errorf("Result must be %s, but got %s instead", tc.result, res)
+			res, err := timeago.Parse(tc.date)
+
+			if err != nil {
+				test.Errorf("Error must be nil, but got %v instead", err)
+			}
+
+			if res != tc.res {
+				test.Errorf("Result must be %s, but got %s instead", tc.res, res)
 			}
 		})
 	}
@@ -66,8 +72,8 @@ func TestParseEn(t *testing.T) {
 
 func TestParseEnWithSeconds(t *testing.T) {
 	cases := []struct {
-		date   time.Time
-		result []string
+		date time.Time
+		res  []string
 	}{
 		{subSeconds(0), []string{"0 seconds ago", "1 second ago"}},
 		{subSeconds(1), []string{"1 second ago", "2 seconds ago"}},
@@ -86,8 +92,14 @@ func TestParseEnWithSeconds(t *testing.T) {
 		t.Run("result for "+tc.date.String(), func(test *testing.T) {
 			timeago.Configure(&timeago.Config{Language: langEn})
 
-			if res := timeago.Parse(tc.date); res != tc.result[0] && res != tc.result[1] {
-				test.Errorf("Result must be %s or %s, but got %s instead", tc.result[0], tc.result[1], res)
+			res, err := timeago.Parse(tc.date)
+
+			if err != nil {
+				test.Errorf("Error must be nil, but got %v instead", err)
+			}
+
+			if res != tc.res[0] && res != tc.res[1] {
+				test.Errorf("Result must be %s or %s, but got %s instead", tc.res[0], tc.res[1], res)
 			}
 		})
 	}
