+++ tests/multiple_options_test.go
@@ -9,13 +9,13 @@ import (
 
 func TestParseWithMultipleFlags(t *testing.T) {
 	type TestCase struct {
-		date   time.Time
-		result string
+		date time.Time
+		res  string
 	}
 
 	cases := []struct {
-		date   time.Time
-		result string
+		date time.Time
+		res  string
 	}{
 		{subMinutes(1), "1 minute"},
 		{subMinutes(5), "5 minutes"},
@@ -47,8 +47,14 @@ func TestParseWithMultipleFlags(t *testing.T) {
 		t.Run("result for "+tc.date.String(), func(test *testing.T) {
 			timeago.Configure(&timeago.Config{Language: langEn})
 
-			if res := timeago.Parse(tc.date, "online", "noSuffix"); res != tc.result {
-				test.Errorf("Result must be %s, but got %s instead", tc.result, res)
+			res, err := timeago.Parse(tc.date, "online", "noSuffix")
+
+			if err != nil {
+				test.Errorf("Error must be nil, but got %v instead", err)
+			}
+
+			if res != tc.res {
+				test.Errorf("Result must be %s, but got %s instead", tc.res, res)
 			}
 		})
 	}
