+++ tests/no_suffix_test.go
@@ -9,9 +9,9 @@ import (
 
 func TestParseWithNoSuffixFlag(t *testing.T) {
 	cases := []struct {
-		date   time.Time
-		result string
-		lang   string
+		date time.Time
+		res  string
+		lang string
 	}{
 		{subMinutes(1), "1 minute", "en"},
 		{subMinutes(2), "2 minutes", "en"},
@@ -67,8 +67,14 @@ func TestParseWithNoSuffixFlag(t *testing.T) {
 		t.Run("result for "+tc.date.String(), func(test *testing.T) {
 			timeago.Configure(&timeago.Config{Language: tc.lang})
 
-			if res := timeago.Parse(tc.date, "noSuffix"); res != tc.result {
-				test.Errorf("Result must be %s, but got %s instead", tc.result, res)
+			res, err := timeago.Parse(tc.date, "noSuffix")
+
+			if err != nil {
+				test.Errorf("Error must be nil, but got %v instead", err)
+			}
+
+			if res != tc.res {
+				test.Errorf("Result must be %s, but got %s instead", tc.res, res)
 			}
 		})
 	}
