+++ tests/ru_test.go
@@ -11,8 +11,8 @@ const langRu = "ru"
 
 func TestParseRu(t *testing.T) {
 	cases := []struct {
-		date   time.Time
-		result string
+		date time.Time
+		res  string
 	}{
 		{subMinutes(1), "1 минута назад"},
 		{subMinutes(2), "2 минуты назад"},
@@ -58,8 +58,14 @@ func TestParseRu(t *testing.T) {
 		t.Run("result for "+tc.date.String(), func(test *testing.T) {
 			timeago.Configure(&timeago.Config{Language: langRu})
 
-			if res := timeago.Parse(tc.date); res != tc.result {
-				test.Errorf("Result must be %s, but got %s instead", tc.result, res)
+			res, err := timeago.Parse(tc.date)
+
+			if err != nil {
+				test.Errorf("Error must be nil, but got %v instead", err)
+			}
+
+			if res != tc.res {
+				test.Errorf("Result must be %s, but got %s instead", tc.res, res)
 			}
 		})
 	}
@@ -67,8 +73,8 @@ func TestParseRu(t *testing.T) {
 
 func TestParseRuWithSeconds(t *testing.T) {
 	cases := []struct {
-		date   time.Time
-		result []string
+		date time.Time
+		res  []string
 	}{
 		{subSeconds(0), []string{"0 секунд назад", "1 секунда назад"}},
 		{subSeconds(1), []string{"1 секунда назад", "2 секунды назад"}},
@@ -88,8 +94,14 @@ func TestParseRuWithSeconds(t *testing.T) {
 		t.Run("result for "+tc.date.String(), func(test *testing.T) {
 			timeago.Configure(&timeago.Config{Language: langRu})
 
-			if res := timeago.Parse(tc.date); res != tc.result[0] && res != tc.result[1] {
-				test.Errorf("Result must be %s or %s, but got %s instead", tc.result[0], tc.result[1], res)
+			res, err := timeago.Parse(tc.date)
+
+			if err != nil {
+				test.Errorf("Error must be nil, but got %v instead", err)
+			}
+
+			if res != tc.res[0] && res != tc.res[1] {
+				test.Errorf("Result must be %s or %s, but got %s instead", tc.res[0], tc.res[1], res)
 			}
 		})
 	}
